Add avro

+++ src/main/java/com/test/processing/models/MarketCap.java
@@ -6,5 +6,5 @@ import java.math.BigDecimal;
 
 @Value
 public class MarketCap {
-    BigDecimal marketCapDifference;
+    double marketCapDifference;
 }
